Feature/milestone 05 (#8)
* Phase A

* Add migration

* Phase H+B - 01

* Phase H+B 02

* Phase H+B 03

* Phase H+B 04 Fix Seeding

* Phase C1

* Phase C2

* Phase E1

* Refactor ApplicationStatus enum

* Refactor FieldType enum

* Phase D1

* Phase D2

* refactor PermitType DTOs

* Move queries + fix PermitType response

* Phase D3

* D4

* D5

* Refactor ReviewDecision Enum

* D6

* D7

* D8

* fix issues after full review

* Phase F

* Phase G

* Blazor Authentication Hardening

* Fix PR build

* Upgrade Identity packages

* Add CODEOWNERS

* Add dependency review action

* Potential fix for pull request finding 'CodeQL / Workflow does not contain permissions'


---------

+++ tests/ATLAS.Application.Tests/Queries/GetApplicationByIdQueryHandlerTests.cs
@@ -3,7 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using ATLAS.Application.DTOs;
 using ATLAS.Application.Interfaces;
-using ATLAS.Application.Queries;
+using ATLAS.Application.Queries.Applications;
 using ATLAS.Domain.Entities;
 using ATLAS.Domain.Enums;
 using ATLAS.Domain.Interfaces;
@@ -53,7 +53,7 @@ namespace ATLAS.Application.Tests.Queries
             Assert.NotNull(result);
             Assert.Equal(application.Id, result.Id);
             Assert.False(string.IsNullOrEmpty(result.ApplicationNumber));
-            Assert.Equal((int)ApplicationStatus.Submitted, result.Status);
+            Assert.Equal(ApplicationStatus.Submitted, result.Status);
         }
 
         [Fact]
